Using props for component data, password incapsulated.
