+++ package.json
@@ -68,7 +68,6 @@
         "react-slick": "^0.23.1",
         "redux": "^4.0.0",
         "redux-saga": "^0.16.0",
-        "redux-thunk": "^2.3.0",
         "resolve": "1.6.0",
         "sass-loader": "^7.1.0",
         "semantic-ui-calendar-react": "^0.8.0",
