+++ src/client/logic/reducer.js
@@ -1,6 +1,7 @@
 import { combineReducers } from 'redux';
 import search from 'client/logic/search/reducer';
 import registration from 'client/logic/registration/reducer';
+import { reducer as formReducer } from 'redux-form';
 
 export default combineReducers({
     search,
